renamed client signup file
