+++ lib/plugins.js
@@ -5,7 +5,7 @@ const Inert 		 = require('inert');
 const Candidate 		 = require('./plugins/candidate.js');
 const Client				 = require('./plugins/client.js');
 const Login 				 = require('./plugins/login.js');
-const ClientSignUp   = require('./plugins/clientSignup.js');
+const ClientSignUp   = require('./plugins/clientsignup.js');
 const Home 					 = require('./plugins/home.js');
 const AdminDashboard = require('./plugins/admindashboard.js');
 const Logout			   = require('./plugins/logout.js');
