+++ lib/plugins/clientsignup.js
